refactor: (scheduler) modularize handlers and search pipelines (#1004)
* refactor(scheduler): modularize handlers and search

- extract scheduler handlers into dedicated modules

- split retriever into pipelines (search/enhance/rerank/filter)

- centralize text search logic for API and scheduler

Refs #1003

* fix: resolve ruff lint errors and address Copilot review feedback

- Fix TC001/TC002/TC003: move type-only imports into TYPE_CHECKING blocks
- Fix RUF059: prefix unused variables with underscore
- Fix typos: "Memorires" -> "Memories", "exeption" -> "exception"
- Remove self-assignment: `text_mem_base = text_mem_base`
- Remove unused `user_context` param from `build_search_context`
- Restore original `QUERY_TASK_LABEL` in activation memory update
- Apply ruff format to all modified files

* fix(redis): serialize schedule messages for streams

- json-encode list/dict fields for Redis XADD

- decode chat_history safely when reading from streams

* fix(examples): use scheduler handlers

- avoid private _memory_update_consumer call

- delegate mem update handler to built-in handler

* style(ruff): fix isinstance union syntax

- apply X | Y form to satisfy UP038

* style(ruff): format message schemas

- align datetime line breaks with ruff format

* refactor: integrate architectural improvements from refactor-scheduler-stage2
This commit merges key modularization benefits and bug fixes from the
refactor-scheduler-stage2 branch into fancy-scheduler:
Modularize activation memory logic into ActivationMemoryManager
Introduce SchedulerSearchService for unified memory search coordination
Extract filtering and reranking logic into MemoryPostProcessor
Maintain intentional search scope of LongTermMemory and UserMemory in SearchPipeline and SchedulerSearchService
Update BaseScheduler to initialize and manage lifecycle of new modules
Refactor BaseSchedulerMemoryMixin to delegate tasks to specialized managers

* refactor: extract common logic to BaseSchedulerHandler and fix import paths

* refactor: rename ctx to scheduler_context in mem_scheduler

Rename abbreviation 'ctx' to 'scheduler_context' in GeneralScheduler and SchedulerHandlerRegistry to improve code readability and clarity.

* refactor: sync orchestrator config and fix scheduler imports

- Sync orchestrator config removal when unregistering handlers in dispatcher
- Fix missing TaskPriorityLevel import in dispatcher
- Fix register_handlers signature in BaseSchedulerQueueMixin
- Fix handler registry imports and initialization map
- Fix relative imports in handlers package

* fix: resolve PR #1 review issues (P0 imports, P3 types/init)

* chore(ruff): fix unused unpacked vars

- prefix unused unpacked vars with underscore

- apply ruff format changes

* perf(search): include embeddings for mmr

* fix: Pass user_context in mem_read and pref_add handlers

- Update MemReadMessageHandler to extract user_context from message and pass it to _process_memories_with_reader and transfer_mem.
- Update PrefAddMessageHandler to extract user_context from message and pass it to pref_mem.add.
- This ensures user context information is available during memory reading and preference adding operations.

---------
