+++ src/memos/mem_scheduler/optimized_scheduler.py
@@ -19,6 +19,7 @@ from memos.mem_scheduler.utils.api_utils import format_textual_memory_item
 from memos.mem_scheduler.utils.db_utils import get_utc_now
 from memos.mem_scheduler.utils.misc_utils import group_messages_by_user_and_mem_cube
 from memos.memories.textual.tree import TextualMemoryItem, TreeTextMemory
+from memos.search import build_search_context, search_text_memories
 from memos.types import (
     MemCubeID,
     SearchMode,
@@ -104,29 +105,14 @@ class OptimizedScheduler(GeneralScheduler):
         mem_cube: NaiveMemCube,
         mode: SearchMode,
     ):
-        """Fine search memories function copied from server_router to avoid circular import"""
-        target_session_id = search_req.session_id
-        if not target_session_id:
-            target_session_id = "default_session"
-        search_priority = {"session_id": search_req.session_id} if search_req.session_id else None
-        search_filter = search_req.filter
-
-        # Create MemCube and perform search
-        search_results = mem_cube.text_mem.search(
-            query=search_req.query,
-            user_name=user_context.mem_cube_id,
-            top_k=search_req.top_k,
+        """Shared text-memory search via centralized search service."""
+        return search_text_memories(
+            text_mem=mem_cube.text_mem,
+            search_req=search_req,
+            user_context=user_context,
             mode=mode,
-            manual_close_internet=not search_req.internet_search,
-            search_filter=search_filter,
-            search_priority=search_priority,
-            info={
-                "user_id": search_req.user_id,
-                "session_id": target_session_id,
-                "chat_history": search_req.chat_history,
-            },
+            include_embedding=(search_req.dedup == "mmr"),
         )
-        return search_results
 
     def mix_search_memories(
         self,
@@ -157,19 +143,13 @@ class OptimizedScheduler(GeneralScheduler):
             ]
 
         # Get mem_cube for fast search
-        target_session_id = search_req.session_id
-        if not target_session_id:
-            target_session_id = "default_session"
-        search_priority = {"session_id": search_req.session_id} if search_req.session_id else None
-        search_filter = search_req.filter
+        search_ctx = build_search_context(search_req=search_req)
+        search_priority = search_ctx.search_priority
+        search_filter = search_ctx.search_filter
 
         # Rerank Memories - reranker expects TextualMemoryItem objects
 
-        info = {
-            "user_id": search_req.user_id,
-            "session_id": target_session_id,
-            "chat_history": search_req.chat_history,
-        }
+        info = search_ctx.info
 
         raw_retrieved_memories = self.searcher.retrieve(
             query=search_req.query,
