+++ src/memos/mem_scheduler/task_schedule_modules/orchestrator.py
@@ -17,11 +17,8 @@ from __future__ import annotations
 
 from memos.log import get_logger
 from memos.mem_scheduler.schemas.task_schemas import (
-    ADD_TASK_LABEL,
-    ANSWER_TASK_LABEL,
     DEFAULT_PENDING_CLAIM_MIN_IDLE_MS,
     PREF_ADD_TASK_LABEL,
-    QUERY_TASK_LABEL,
     TaskPriorityLevel,
 )
 from memos.mem_scheduler.webservice_modules.redis_service import RedisSchedulerModule
@@ -38,11 +35,7 @@ class SchedulerOrchestrator(RedisSchedulerModule):
         """
         # Cache of fetched messages grouped by (user_id, mem_cube_id, task_label)
         self._cache = None
-        self.tasks_priorities = {
-            ADD_TASK_LABEL: TaskPriorityLevel.LEVEL_1,
-            QUERY_TASK_LABEL: TaskPriorityLevel.LEVEL_1,
-            ANSWER_TASK_LABEL: TaskPriorityLevel.LEVEL_1,
-        }
+        self.tasks_priorities = {}
 
         # Per-task minimum idle time (ms) before claiming pending messages
         # Default fallback handled in `get_task_idle_min`.
@@ -53,6 +46,37 @@ class SchedulerOrchestrator(RedisSchedulerModule):
 
     def get_stream_priorities(self) -> None | dict:
         return None
+
+    def set_task_config(
+        self,
+        task_label: str,
+        priority: TaskPriorityLevel | None = None,
+        min_idle_ms: int | None = None,
+    ):
+        """
+        Dynamically register or update task configuration.
+
+        Args:
+            task_label: The label of the task.
+            priority: The priority level of the task.
+            min_idle_ms: The minimum idle time (ms) for claiming pending messages.
+        """
+        if priority is not None:
+            self.tasks_priorities[task_label] = priority
+        if min_idle_ms is not None:
+            self.tasks_min_idle_ms[task_label] = min_idle_ms
+
+    def remove_task_config(self, task_label: str):
+        """
+        Remove task configuration for a specific label.
+
+        Args:
+            task_label: The label of the task to remove configuration for.
+        """
+        if task_label in self.tasks_priorities:
+            del self.tasks_priorities[task_label]
+        if task_label in self.tasks_min_idle_ms:
+            del self.tasks_min_idle_ms[task_label]
 
     def get_task_priority(self, task_label: str):
         return self.tasks_priorities.get(task_label, TaskPriorityLevel.LEVEL_3)
