+++ src/memos/search/__init__.py
@@ -1,0 +1,4 @@
+from .search_service import SearchContext, build_search_context, search_text_memories
+
+
+__all__ = ["SearchContext", "build_search_context", "search_text_memories"]
