+++ src/memos/types/openai_chat_completion_types/__init__.py
@@ -1,4 +1,4 @@
-# ruff: noqa: F403, F401
+# ruff: noqa: F403
 
 from .chat_completion_assistant_message_param import *
 from .chat_completion_content_part_image_param import *
